requirements.txt
fastapi==0.103.1
pydantic==2.3.0
uvicorn==0.23.2
redis[hiredis]==5.0.0
python-jose==3.3.0
